Revamp disassembly tracing
DisasmFind is now DisasmTrace.

Instead of returning a single address, it now takes up to 16 symbol
outputs that can be filled in through matching the pattern. A single
trace may fill out many symbols at once.

This reduces duplication as well as encourages focus on well-developed
high quality trace patterns, rather than spamming many quick and
dirty copies.

Tracing now has the ability to backtrack and try for a different match
by skipping more instructions (up to the current skip max). This
improves the ability to adapt to multiple code versions with robust
checks that reduce false negatives / positives.

Reworked trace op structure to make it less verbose and easier to
read and write. Pseudo-ops are specified using macros.

Add LABEL and GOTO pseudo-ops that allow storing the current EIP
and returning to it later. This is useful for a trace to be able to
capture outputs from both branches of a conditional jump.

Improved symbol logic to distinguish between a symbol that has not yet
been resolved (or failed to resolve) and one that did resolve but
contains a 0 value.