Add ARG_DUMMY type to hookgen / wrapgen
This is a special argument that the native program takes, but the
wrapped / replaced function does not.

This is mostly useful for if there are different versions of the
function, and which one is to be hooked is not known until runtime.

In this case, multiple FuncInfo structures can be provided to work
with the same native function or call and adapt to the extra
argument (which is often unneeded).