Reorganize source tree a bit
Move all of the hooks under hook/

Rename the hook source files to include class and method, to be
cognizant that we're using the VS generator on Windows and having
source files wih the same name results in clashes because it puts all
of the object files in the same directory, regardless of where they
are in the source tree.

Move all of the ftl module headers to ftl/ and consolidate with the
mostly redundant function_*.h headers

Shorten the name of the OS/Arch-specific source files with game function
info and traces from functions_*.c to ftl_*.c.