Move minicrt's snprintf to xsnprintf
Move it out of the way because it can interfere with windows headers,
and because it's lacking some key features like floating point numbers.

Use as x*printf early on before the module is loaded, then use the
MSVCRT wrapper after that.

Rename ftl_mem to ftl_stdlib since it includes more than just malloc
and free now.