Rework how SubspaceUI reference handling works
UI thread now acquires and holds a reference to the UI object while
it is running.

This means UI objects don't need to go through the trouble of
holding / converting weak references. They can just keep a pointer to
the UI object as they're guaranteed that the SubspaceUI object will
have a longer lifetime than they will.

Standardize on IUP property names for storing object references, and
add some convenience macros for it.