Don't use v141_xp toolchain for subspacegame
Originally I set it for that so that it could link to msvcrt.dll,
because mingw32-compiled FTL links against it.

However it's not actually necessary, because the game client is compiled
as a bare executable with /NODEFAULTLIB, so it doesn't get any C runtime
at all anyway. None of the special handling that v141_xp brings in is
actually needed.

Changing this allows for the use of the ninja generator and brings it in
line with the main Subspace build. It should also make automated builds
on GitHub easier as it can use a modern compiler version for all of the
pieces.