Start adding waypoints
These are special symbols that do not correspond to any real
function or variable, but rather high-traffic and especially useful
points in the code that can be used to find a number of other
symbols.

The first waypoint is a switch() statement in CApp::OnLoop that
dispatches commands from CommandGui. Break up a single monolithic
trace of OnLoop into smaller pieces, so if there are version
incompatibilies with one of them, the others may still succeed.