Add a missing required symbol to timewarp code
Also add a pseudo-assert in debug builds; set an invalid handle to
module analysis data once it's been freed. This causes a crash if
anything tries to resolve a symbol after the initial load. This should
help ensure there aren't any other symbols that get resolved late.

Releaes builds still rebuild the analysis data if needed, causing the
game to freeze for a second or so while that's done.