DisasmTrace: Fix LABEL after SKIP
Since skip just sets state variables, using these two together was very
unintuitive as it meant the label would be created with the address
that the skip started.

Change it to also set a state variable, so the label applies to the
first matched instruction after the skip, which is almost always what
we want.