Switch (back) to old codegen allocation scheme
Rather than using specialized a smalloc heap, use the original growing
array of pages. This has some advantages.

- Only commits a single page at a time instead of a large block.
- Keeps the generated code in proper R-X pages instead of writable most
  of the time; only switchtes to RWX when generating new wrappers.
- Can tightly pack the generated functions together; smalloc
  theoretically could have done this had I bothered to implement
  block shrinking, but this lets me keep it simpler.
- Not really a downside that the memory area can only grow -- once
  wrappers are generated they are never deallocated anyway.

It does have a theoretical upper limit since it only reserves 1024
pages (4MB), but so far all of the generated wrappers fit in a single
page so that is not going to be hit, probably ever.