Refine savefile compat tweak
Instead of deleting the save and renaming the new one into place,
rename the old one first.

This is to catch a corner case where some other process has
continue.sav open with a read lock, so the file cannot be deleted,
which would also prevent the new file from being renamed.

The process is now:
- Rename continue.sav to continue.sav.subspace-old
- Rename continue.sav.subspace-new to continue.sav
- Delete continue.sav.subspace-old

The old file is also deleted if present later; in case it was unable
to be deleted right away.