Add UI string cache
The ephemeral C strings returned by strC are not always appropriate
for use with IUP, because some attributes do not copy the string but
only keep a pointer to it.

The string cache is just an object associated with a UI object that
keeps a non-rope copy of any strings that IUP may be referencing, and
automatically cleans them up when the object is destroyed.