Remove UIStrCache
It's not actually needed because IUP already has a built-in way of
handling this situation. Derp.

Rework langGetC to take a C string key and guarantee it always returns
a fixed pointer that will not change.

For anything dynamically generated, use IupSetStrAttribute, which
internally duplicates the string in the hash table and avoids the
stale pointer problem.