Rework current run
Current run focused by the UI should not be a child of the GameInst,
because we want to be able to load historical data from the db, as well
as continue to display the last run's stats even after the game exits.

Split this into a RunInfo that is owened by subspace itself (for the
UI), and an active run owned by the GameInst, so that it can continue
to record data even if a different run is loaded for display.