Initial implementation of run logging
Run log and combat log entries are saved to the database, and
sent to UI elements interested in them.

Still a work in progress but for now enables actual scrap counting
in the run info (after SRA / Repair Arm) as well as very basic first
draft of the scrap per sector graph.