Fix crash / graphical bug when splitting panels
Crash was caused by an IUP bug. When a matrix is unmapped, it freed
the memory used for merge data, but did not reset the internal
attributes used to track merge state.

Graphical bug caused by matrix losing its contents when unmapped. Added
a mechanism to notify panels that an unmap/remap has occured and they
need to reset the state of any controls that lose data when unmapped.