Add priority system for outbound control messages
Only for the game portion, this system helps ease the pain of not
always being able to choose the best place to hook and having to deal
with existing control flow.

Instead of outbound control messages being sent immediately, they
are batched until the end of the frame. A priority number can be set;
lower priorities are sent first. This alows multiple messages
generated on the same frame to be kept in the correct order.