Rework timestamps
No longer send OS-specific timestamps over the control protocol.

Instead of discrete fields, the CX timestamp (which uses frametime
for messages originating from the game) is now part of the standard
message header.

Add some support functions on the Win32 build for doing the 64-bit
math necessary for this without the standard library functions that
are normally implicitly called for 64-bit division on 32-bit CPUs.