Fix some oddities around Sector runlog event
Because of when the hook is called, it was using the previous savepoint
(number of beacons have not been incremented), as well as the gametime
from the last beacon. Go ahead and reset gametime when advancing to
a new sector as well, and assume that the beacon count is one higher
since it's about to be incremented.