Rework log relay & add parallel UI notify system
Rework log relay to allow the subscriber to specify that it wants
the callback to be dispatched to the UI thread instead of the main
pool.

Ensure that the reset / replay complete callbacks are called in the
correct order and from the correct thread.

The UI notify system is similar, but allows for freeform event
messages with arbitrary parameters.