Add Scraptracker
Start building data models that are part of the run info so they can
be used by multiple UI elements instead of having the UI track all
the state itself.

First is the scrap tracker, which aggregates scrap gains and losses.

Rewrite scrap graph panel to use UI notifications sent by the
scrap tracker to know when to update the displayed data.