Fix save files not being created for some jumps
It seems that PrepareAutoSave is not called for some beacons, notably
revisiting the federation base.

Switch back to the original implementation of performing the save
in the CApp loop. This save happens a bit earlier than the normal
autosave does (before the jump-in animation finishes), but upon
testing seems to cause no obvious ill effects. Something to keep an
eye on if it causes issues.