Auto-abandon restarted runs
Properly mark runs abandoned when they are being actively tracked and
the player uses the 'Restart' or 'Hangar' menu options to permanently
quit.

This is not able to catch the case of using 'New Game' from the main
menu to delete an existing save, but Subspace has no real way to know
WHICH savegame is being deleted and can't guarantee the player hasn't
shuffled the save file around, so safer to just not handle that case.