Fix occasional crash loading practice saves
For some as yet unknown reason the game sometimes randomly crashes
when loading a practice save if a previous practive save has been
loaded.

Adding a 1 frame pause between returning to the menu and loading the
save seems to fix this; probably some extra cleanup is being done
in the main loop.

This frame is hidden and not rendered to avoid a flicker.