Start implementing screenshot feature
First part is functional -- saving the screen data to a raw file to be
later compressed by the main process.

There are some imporant limitations that need to be documented. This
only works when the framebuffer is being used; that is in windowed mode
or fullscreen stretched (not native or bars).

It also does not work with the D3D renderer; seemingly due to bugs
in the SIL library but further examination is needed to see if they
can be fixed.