Raw screenshot functionality in all scenarios
Work around the various issues. When using D3D or not in framebuffer
mode, use a different method to capture the contents of the back
buffer directly. This also makes it work with bordered fullscreen
and native fullscreen.

One small issue remains. In scaled (windowed or fullscreen) mode
on D3D the scaled image is screenshotted rather than the native
resolution like it is with the OpenGL renderer.