Link subspacegame against static C runtime
The project that used the prototype harness that was re-used and
extended for Subspace made extensive use of TLS and did not function
correctly if the C runtime was already initialized by the loader
process.

This seems to not be an issue for FTL, either because the game does not
make use of TLS, or because it's using an old enough runtime that
it doesn't conflict when subspacegame statically links in the current
version.

In any case, this seems to be working. If it proves to be stable, it
will enable the use of the cx library in the main game portion,
vastly simplify data structure management, and allow for the removal
of minicrt and the suboptimal allocator it provides.